In addition to cosmetic and logfile related changes, this fixes a real bug in the network recv() call which had the MSG_WAITALL flag, which caused a hang on the last (partially filled buffer) recv of a file transfer.
Also, the "OK" message from the server to the client, now has in it the
connection number from the server, the output file name from the server,
and the number of bytes written.  This should make it easy to verify that
all bytes of a file are identical in size and content after a file transfer.